+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Version 0.1.10 (February 24th 2019)
 -----------------------------
- * Added base64 encoding option
- * Added ability to set headers into HTTP requests methods
+ * Improved logger messages
+ * Updated Mouette.js dependency to version 0.2.9
  
 Version 0.1.9 (December 24th 2018)
 -----------------------------